- compile time fixes for MacOSx - removed all <SDL/SDL*.h> includes - fixed search path for Font Loading
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@866 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/immovable.cc
@@ -794,7 +794,7 @@ Cleanup
 PlayerImmovable::~PlayerImmovable()
 {
 	if (m_workers.size())
-		log("Building::~Building: %i workers left!\n", m_workers.size());
+		log("Building::~Building: %i workers left!\n", (int)m_workers.size());
 }
 
 /*
